Soak passes; move draw snapshots off the UI stack, fix redraw overflow
8 hours of back-to-back sweeps on production 7a0b70c, WiFi off then on, plus
7.6 hours of accidental idle tail after USB was pulled and the board kept
running on battery. 4,112 laps, 0 failures, home restored every lap, every drop
and error counter zero.

Bounded memory is confirmed, and the accidental tail is what confirms it: free
heap moved in exactly two discrete steps (UI canvas at 0.02h, WiFi AP at 4.02h)
and then sat at 153,928 bytes unchanged for 10.7 hours with heap_largest flat
beside it. A leak would have shown in a window where nothing was allocating.

Three findings that were not what the run was looking for:

Pass-B spent 3.90h of the 8 and promoted nothing (PBA=15680, PBD=0). Sweep away
time is 80.2% Pass-B against 19.8% Pass-A: a 0-peak lap costs 833ms and each
peak adds 6-11s. At 60.8% radio-away, phase12 §6.3's proportionality implies
~2,650 packets not heard for zero promotions. Not a defect — Pass-B promotes
off-grid packets and the traffic here was on the home channel — but it is the
measured cost half of a question Workstream 17 exists to ask.

The UI task was at 93% of its 4096-byte stack: ui_stack_free read 288 bytes, a
watermark reached in the first two minutes and never moved. -fstack-usage on the
shipped build traced it to snapshot structs held as stack locals in draw
functions — CaptureHistory 516B, NodeRoster 672B, ScopeTrace 256B, WaterfallRow
236B. Only ui_task ever draws, so all eleven are now function-static: BSS
instead of stack. Worst draw chain 1,312B -> 592B, RAM 18.3% -> 19.8%. Worth
being clear that 288B was the deepest path actually exercised — the device sat
on one card for most of the run, so pages never displayed contributed nothing,
and the real margin was thinner than the number.

The redraw instrumentation added the day before overflowed. uiRedrawTotalUs was
a uint32 of microseconds accumulating ~161,200 us/s at the measured 2.6 fps, so
it wrapped at 7.4h and the reported mean fell from 62,781 to 5,306 at the 8-hour
mark. Now uint64, with frame count exposed so a host can difference samples for
a windowed mean. The max was unaffected — 223ms worst frame, last increased at
4.92h and flat through the remaining 10.7h.

Not flashed: the board was disconnected to read the card. The open follow-up is
re-checking ui_stack_free after deliberately walking every card view, which this
run could not measure.

259/259 host tests pass.

+++ src/ui_task.cpp
@@ -562,7 +562,12 @@ KeyAction pollKeyAction() {
 // and a per-page table would cost more RAM than the answer is worth. Serial
 // Control's STATUS carries it, so a fixture can sample it without a display.
 volatile uint32_t uiRedrawMaxUs = 0;
-volatile uint32_t uiRedrawTotalUs = 0;
+// uint64: at the measured ~2.6 fps and ~62 ms mean, a uint32 of microseconds
+// wraps in about 7.4 hours — run0089 caught it doing exactly that just before
+// the 8-hour mark, where the reported mean fell from 62,781 to 5,306 us. The
+// max is a running maximum and was unaffected; only the mean was wrong, and
+// only after the wrap.
+volatile uint64_t uiRedrawTotalUs = 0;
 volatile uint32_t uiRedrawCount = 0;
 
 void fullRedraw() {
@@ -955,6 +960,9 @@ uint32_t uiRedrawMeanUs() {
     const uint32_t n = uiRedrawCount;
     return n ? (uint32_t)(uiRedrawTotalUs / n) : 0;
 }
+// Frame count is exposed so a host can tell a genuine mean from one computed
+// over too few frames, and can difference two samples for a windowed mean
+// rather than reading the since-boot average.
 uint32_t uiRedrawFrames() { return uiRedrawCount; }
 void uiRedrawStatsReset() {
     uiRedrawMaxUs = 0;
